requirements.txt
torch
torchvision
pytorch-lightning
hydra-core
cv2
albumentations
tensorboard
